Fix: order of parameter in `split` function. The `merge` relation can be illustrated as follow:
    merge xs ys zs ≡ xs + ys = zs

This shows a key point: every element of `xs` is contained in `zs`.

The original definition of `split` makes no sense under this revelation.

    split : ∀ {A : Set} {P : A → Set} (P? : Decidable P) (xs : List A)
      → ∃[ ys ] ∃[ zs ] ( merge xs ys zs × All P ys × All (¬_ ∘ P) zs )

`zs` will contain all elements of `xs` (the input!) and we also have that `All (¬_ ∘ P) zs`. This function signature cannot be inhabited, as it would imply that for any `P : A → Set` and any `x : A` we have `¬(P x)`.