Prettify renamings by defining subset relation for contexts (`_⊆_`). Considering that the book tries to make every formula simple to read, I believe that the definition of this small auxiliary function is well worthwhile.
I would also like to define another function for substitutions, but I cannot think of a suitable name for it. In my notes, I called it `_⊑_` to give a hint about how it resembles `_⊆_`. An example of its use can be seen here:

https://gist.github.com/pedrominicz/27c27aa491f981b6c2eebd64f5137348

I believe I have seen functions similar, if not equal, to `_⊑_` defined somewhere in the literature, but I could not find it. Name suggestions are welcome.