EPUB: Populate acknowledgements.md with contributors
Use liquid-lua to generate acknowledgements_epub.md, which is used by
pandoc to build the EPUB book.  Also stop using the '{%-' / '-%}'
syntax, which liquid-lua does not support, in templates.

Also add a vertical space between the first bullet in each bulleted
list so pandoc recognizes bulleted lists as such. This has no effect on
the the website.