Fix remaining epubcheck warnings
HTML does not permit nested <p> elements, which epubcheck complains
about. This change fixes those remaining warnings.

This change also causes the Dedication on the website to be centered
when viewed in Safari, which it was not before. The EPUB output looks
the same in either version with Thorium reader 1.3 on macOS.