remove paragraph that duplicates what's in the exercises
The presence of this paragraph here suggests that this approach should
be used for <-trans.  So that's what I did.  But then when I read
further I saw that this paragraph is duplicated in the ≤-iff-< and
<-trans-revisited exercises, and actually this approach is intended
for those exercises.  So I was a bit confused.

I don't know what the appropriate wording/ordering is here, but I
think just deleting this paragraph would be reasonable since the
content is duplicated in the ≤-iff-< and <-trans-revisited exercises.