move postulate of foldr-++ up to the exercise and generalize it
This follows the pattern set by the previous chapters, which always
postulate things proved by exercises and needed by later proofs, in
the exercise itself.  It's also more convenient when doing the
exercise - one can just remove the "postulate" as a purely local
change, rather than having to scroll down and delete the postulate
from where it's used later.

Plus, the postulate of foldr-++ before took a _⊗_ : A → A → A, which
is insufficiently general for all the exercises, namely foldr-∷.