Bisimulation: environment -> context
In Bisimulation, there are two occurrences of "environment", I believe
the correct term is "context".
In the context of PL theory in general, they are interchangable.
However, I think for it is good to be pedantic in a textbook, and that
it's good to disambiguate that from "environments" in the BigStep
chapter.

Signed-off-by: Fangyi Zhou <me@fangyi.io>